refactor(tests): extract ValidateCoverageStepDefinition from CorrectUsageStepDefinitions
